--- react-todo-app/README.md
@@ -1,60 +1,0 @@
-![](https://img.shields.io/badge/Microverse-blueviolet)
-
-# React-Todo-App
-
-> A React based Todo-App based on the ReactJS [tutorial](https://ibaslogic.com/react-tutorial-for-beginners/)
-
-
-## Built With
-
-- HTML, CSS, JavaScript
-- React JS
-
-## Live Demo 
-
-[Live demo](https://jaymelfah.github.io/React-todo-app/)
-
-
-## Getting Started
-
-This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
-
-
-To get a local copy up and running follow these simple example steps.
-
-### Install
-Have NodeJS installed and follow the next steps
-### Usage
-- `npm i` for installing dev dependecies
-- `npm run build` or `npm start` to see the project
-### Run tests
-- `npm test`
-### Deployment
-- `npm run deploy`
-
-## Author
-
-👤 Jeremiah Melfah
-
-- GitHub: [GitHub](https://github.com/Jaymelfah)
-- LinkedIn: [LinkedIn](https://www.linkedin.com/in/jeremiah-ekow-melfah-a4402a161/)
-
-## 🤝 Contributing
-
-Contributions, issues, and feature requests are welcome!
-
-Feel free to check the [issues page](../../issues/).
-
-## Show your support
-
-Give a ⭐️ if you like this project!
-
-## Acknowledgments
-
-- Hat tip to anyone whose code was used
-- Inspiration
-- etc
-
-## 📝 License
-
-This project is [MIT](./MIT.md) licensed.
