+++ src/.babelrc
@@ -1,0 +1,6 @@
+{
+    "presets": [
+        "@babel/preset-react"
+    ],
+    "plugins": ["@babel/plugin-syntax-jsx"]
+  }
